% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.glinternet.R
\name{cv.glinternet}
\alias{cv.glinternet}
\title{Learn a quadratic regression model via hierarchical group-lasso regularization from \code{glinternet} package.}
\usage{
cv.glinternet(x, y, family = "gaussian", ...)
}
\arguments{
\item{x}{An \code{n} by \code{p} design matrix of main effects. Each row is an observation of \code{p} main effects.}

\item{y}{A response vector of size \code{n}.}

\item{family}{A character string describing the target variable: "gaussian" for continuous (the default), "binomial" for logistic.}

\item{...}{other arguments to be passed to the \code{glinternet} calls.}
}
\value{
An object of S3 class "\code{cv.glinternet}".
 \describe{
  \item{\code{n}}{The sample size.}
  \item{\code{p}}{The number of main effects.}
  \item{\code{family}}{The response \code{family} used in \code{cv.glinternet}.}
  \item{\code{a0}}{Intercept value.}
  \item{\code{compact}}{A compact representation of the selected variables. \code{compact} has three columns, with the first two columns representing the indices of a selected variable (main effects with first index = 0), and the last column representing the estimate of coefficients.}
 }
}
\description{
Learn a quadratic regression model via hierarchical group-lasso regularization from \code{glinternet} package.
}
