% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.hierNet.R
\name{cv.hierNet}
\alias{cv.hierNet}
\title{Build a regression model with hierarchically constrained pairwise interactions from \code{hierNet} package.}
\usage{
cv.hierNet(
  x,
  y,
  family = "gaussian",
  nlam = 20,
  nfolds = 10,
  strong = TRUE,
  ...
)
}
\arguments{
\item{x}{An \code{n} by \code{p} design matrix of main effects. Each row is an observation of \code{p} main effects.}

\item{y}{A response vector of size \code{n}.}

\item{family}{response type. Default is ’gaussian’. The other choice is ’binomial’ for logistic regression}

\item{nlam}{Number of values of lambda to be tried.}

\item{nfolds}{Number of folds in cross-validation. Default value is 10.}

\item{strong}{Indicator to specify strong hierarchy (TRUE) or weak hierarchy (FALSE). Default FALSE.}

\item{...}{other arguments to be passed to the \code{hierNet} calls.}

\item{num_keep}{Number of variables to keep in the screening phase}
}
\value{
An object of S3 class "\code{cv.hierNet}".
 \describe{
  \item{\code{n}}{The sample size.}
  \item{\code{p}}{The number of main effects.}
  \item{\code{family}}{The response \code{family} used in \code{cv.hierNet}.}
  \item{\code{strong}}{The \code{strong} parameter passed into \code{cv.hierNet}.}
  \item{\code{a0}}{Intercept value.}
  \item{\code{compact}}{A compact representation of the selected variables. \code{compact} has three columns, with the first two columns representing the indices of a selected variable (main effects with first index = 0), and the last column representing the estimate of coefficients.}
 }
}
\description{
Build a regression model with hierarchically constrained pairwise interactions from \code{hierNet} package.
}
