% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fields.R
\name{find_capability_mapping}
\alias{find_capability_mapping}
\title{Update the internal mapping of the field containing capability}
\usage{
find_capability_mapping()
}
\value{
Invisibly returns the new value of the capability field
}
\description{
This is a custom field likely only used by the package author. Until
a more flexible method is devised, this helper function finds the
value of the capability field in the Jira schema.
}
\examples{
NULL
}
