% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fields.R
\name{find_field}
\alias{find_field}
\title{Find the ID of a field in the Jira schema, given the field name and type}
\usage{
find_field(fields, name, type)
}
\arguments{
\item{fields}{Output of \code{get_fields()}}

\item{name}{Name of the field to find}

\item{type}{Type of the field to find}
}
\value{
The ID of the field in the schema
}
\description{
Specifying both the name and the type of the field is necessary to
avoid issues with fields that have the same name, but multiple types
defined.
}
\examples{
NULL
}
