% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fields.R
\name{find_story_point_mapping}
\alias{find_story_point_mapping}
\title{Update the internal mapping of the field containing story points}
\usage{
find_story_point_mapping()
}
\value{
Invisibly returns the new value of the story points field
}
\description{
The story point field in Jira unfortunately is mapped to a random
custom field identifier in each instance/project. \code{find_story_point_mapping()}
queries the field list and updates the \code{sprintr} mapping. This value is
not persisted and needs to be re-run each session.
}
\examples{
NULL
}
