% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jira_api.R
\name{get_all_sprints}
\alias{get_all_sprints}
\title{Get sprints associated with a board}
\usage{
get_all_sprints(board_id)
}
\arguments{
\item{board_id}{ID of the origin board to retrieve sprints}
}
\value{
dataframe
}
\description{
Return all of the sprints associated with a specific board. As a project
may have multiple boards, in order to retrieve all sprints in a given
project the user must query all boards and combine the results.
}
\examples{
NULL
}
