% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_boards.R
\name{get_boards}
\alias{get_boards}
\title{Get all available boards}
\usage{
get_boards()
}
\value{
dataframe
}
\description{
Fetches all visible boards in the Jira instance and returns them, and their
metadata, as a dataframe.
}
\examples{
NULL
}
