% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_epic.R
\name{get_epic}
\alias{get_epic}
\title{Get details on a specific epic}
\usage{
get_epic(epic_key)
}
\arguments{
\item{epic_key}{Key of epic to retrieve}
}
\value{
A tibble.
}
\description{
Using an epic_key, get the full details on an epic.
}
\examples{
\dontrun{
get_epic("XXX-1234")
}
}
