% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_epic.R
\name{get_epics}
\alias{get_epics}
\title{Get all available epics on a specific board}
\usage{
get_epics(board_id)
}
\arguments{
\item{board_id}{ID of the origin board to retrieve sprints}
}
\value{
dataframe
}
\description{
Fetches all visible epics in a given Jira board, returning them and their
metadata as a dataframe.
}
\examples{
NULL
}
