% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jira_api.R
\name{get_fields}
\alias{get_fields}
\title{Get all configured fields in a Jira instance}
\usage{
get_fields()
}
\value{
dataframe
}
\description{
Jira contains a large number of both system-supplied and user-configurable
fields. This function fetches the full list of available fields and
their metadata.
}
\examples{
NULL
}
