% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_issue.R
\name{get_issue}
\alias{get_issue}
\title{Get details on a specific issue}
\usage{
get_issue(issue_key, fields = NULL, full_response = FALSE)
}
\arguments{
\item{issue_key}{Key of issue to retrieve}

\item{fields}{Comma separated string of field IDs to retrieve}

\item{full_response}{Return raw list of fields}
}
\value{
A tibble if neither \code{fields} nor \code{full_response} are specified, otherwise a list
}
\description{
Using an issue_key, get the full details on an issue. Defaults to returning a
minimal set of fields as a dataframe, but may also be used to retrieve any
or all available fields.
}
\examples{
\dontrun{
# basic details on an issue as tibble
get_issue("XXX-1234")

# a custom field set as list
get_issue("XXX-1234", fields = 'status,resolution')

# all available fields as list
get_issue("XXX-1234", full_response = TRUE)
}
}
