% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_epic.R
\name{get_issues_by_epic}
\alias{get_issues_by_epic}
\title{Get all issues associated with a given epic}
\usage{
get_issues_by_epic(epic_key)
}
\arguments{
\item{epic_key}{Key of issue to retrieve}
}
\value{
A tibble.
}
\description{
With a given epic key, find all of the associated issues.
}
\examples{
\dontrun{
get_issues_by_epic("XXX-1234")

}
}
