% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_issues_on_backlog}
\alias{get_issues_on_backlog}
\title{Get all backlogged issues.}
\usage{
get_issues_on_backlog(board_id)
}
\arguments{
\item{board_id}{ID of board to retrieve}
}
\value{
dataframe
}
\description{
Retrieves all the issues that are on a board's backlog.
}
\examples{
NULL
}
