% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jira_api.R
\name{get_sprint}
\alias{get_sprint}
\title{Get detailed information on a particular sprint}
\usage{
get_sprint(sprint_id)
}
\arguments{
\item{sprint_id}{ID of the sprint to retrieve}
}
\value{
A tibble
}
\description{
Given a particular sprint ID, retrieve the descriptive information about
that sprint.
}
\examples{
NULL
}
