% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jira_api.R
\name{get_sprint_report}
\alias{get_sprint_report}
\title{Title}
\usage{
get_sprint_report(sprint_id)
}
\arguments{
\item{sprint_id}{ID of sprint to retrieve}
}
\value{
dataframe
}
\description{
Title
}
\examples{
NULL
}
