% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jira_api.R
\name{get_sprint_report_detail}
\alias{get_sprint_report_detail}
\title{Title}
\usage{
get_sprint_report_detail(sprint_id)
}
\arguments{
\item{sprint_id}{ID of the sprint to pull issue-level details on}
}
\value{
A dataframe
}
\description{
Title
}
\examples{
NULL
}
