% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hier_lasso.R
\name{hier_lasso}
\alias{hier_lasso}
\title{Two-stage hierarchical lasso}
\usage{
hier_lasso(
  x,
  y,
  lambda = NULL,
  nlam = 100,
  lam_choice = "min",
  lam_min_ratio = ifelse(nrow(x) < ncol(x), 0.01, 1e-04),
  nfold = 5,
  foldid = NULL
)
}
\arguments{
\item{x}{An \code{n} by \code{p} design matrix of main effects. Each row is an observation of \code{p} main effects.}

\item{y}{A response vector of size \code{n}.}
}
\value{
An object of S3 class "\code{cv.hier}".
 \describe{
  \item{\code{n}}{The sample size.}
  \item{\code{p}}{The number of main effects.}
  \item{\code{fit}}{The whole \code{cv.glmnet} fit object.}
  \item{\code{compact}}{A compact representation of the selected variables. \code{compact} has three columns, with the first two columns representing the indices of a selected variable (main effects with first index = 0), and the last column representing the estimate of coefficients.}
 }
}
\description{
An implementation of the two-stage lasso studied in Hao et, al (2018).
}
\examples{
set.seed(123)
n <- 100
p <- 200
# dense input
x <- matrix(rnorm(n * p), n, p)
y <- x[, 1] - 2 * x[, 2] + 3 * x[, 1] * x[, 3] - 4 * x[, 4] * x[, 5] + rnorm(n)
mod <- hier_lasso(x = x, y = y)

}
