% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jira_api.R
\name{jira_api_post}
\alias{jira_api_post}
\title{Title}
\usage{
jira_api_post(path, post_data)
}
\arguments{
\item{path}{API endpoint to retrieve}

\item{post_data}{List of data to post to API}
}
\value{
A jira_api object
}
\description{
Title
}
\examples{
NULL
}
