% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cv.sprinter.R
\name{plot.cv.sprinter}
\alias{plot.cv.sprinter}
\title{Plot function of cv.sprinter fit}
\usage{
\method{plot}{cv.sprinter}(fit)
}
\arguments{
\item{fit}{A "\code{cv.sprinter}" object.}
}
\description{
This function produces plots of cross-validation for cv.sprinter.
}
\details{
The orange pairs on the top of the plot shows the number of non-zero (main effects, interactions) selected by each value of lambda. Adopted from the function \code{plot.cv.rgam} from package \code{relgam} by Kenneth Tay and Robert Tibshirani.
}
\examples{
set.seed(123)
n <- 100
p <- 200
# dense input
x <- matrix(rnorm(n * p), n, p)
y <- x[, 1] - 2 * x[, 2] + 3 * x[, 1] * x[, 3] - 4 * x[, 4] * x[, 5] + rnorm(n)
mod <- cv.sprinter(x = x, y = y)

plot(mod)
}
\seealso{
\code{\link{cv.sprinter}}.
}
