% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.sprinter.R
\name{plot.sprinter}
\alias{plot.sprinter}
\title{Plot function of sprinter fit}
\usage{
\method{plot}{sprinter}(fit, which = 1, label = TRUE, index = NULL)
}
\arguments{
\item{fit}{Fitted \code{sprinter} object.}

\item{which}{The tuning parameter considered in Step 2.}

\item{label}{If \code{TRUE} (default), annotate the plot with variable labels.}

\item{index}{Lambda indices to plot}
}
\description{
Produces a two-panel plot of the sprinter object showing coefficient paths for both main effects and interactions.
}
\details{
A two panel plot is produced, that summarizes the main effects (left) and interaction (right) coefficients, as a function of lambda. Adopted from the function \code{summary.rgam} from package \code{relgam} by Kenneth Tay and Robert Tibshirani.
}
\examples{
set.seed(123)
n <- 100
p <- 100
# dense input
x <- matrix(rnorm(n * p), n, p)
y <- x[, 1] - 2 * x[, 2] + 3 * x[, 1] * x[, 3] - 4 * x[, 4] * x[, 5] + rnorm(n)
fit <- sprinter(x = x, y = y)

plot(fit)

}
