% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ramp.fit.R
\name{ramp.fit}
\alias{ramp.fit}
\title{Build a quadratic regression model with regularization algorithm under marginality principle from \code{RAMP} package.}
\usage{
ramp.fit(x, y, family = "gaussian", hier = "Strong", ...)
}
\arguments{
\item{x}{An \code{n} by \code{p} design matrix of main effects. Each row is an observation of \code{p} main effects.}

\item{y}{A response vector of size \code{n}.}

\item{family}{response type. Default is ’gaussian’. The other choice is ’binomial’ for logistic regression}

\item{hier}{whether to specify strong or weak heredity. Default is ’Strong’.}

\item{...}{other arguments to be passed to the \code{RAMP} calls.}
}
\value{
An object of S3 class "\code{ramp.fit}".
 \describe{
  \item{\code{n}}{The sample size.}
  \item{\code{p}}{The number of main effects.}
  \item{\code{family}}{The response \code{family} used in \code{ramp.fit}.}
  \item{\code{hier}}{The \code{hier} parameter passed into \code{ramp.fit}.}
  \item{\code{a0}}{Intercept value.}
  \item{\code{compact}}{A compact representation of the selected variables. \code{compact} has three columns, with the first two columns representing the indices of a selected variable (main effects with first index = 0), and the last column representing the estimate of coefficients.}
 }
}
\description{
Build a quadratic regression model with regularization algorithm under marginality principle from \code{RAMP} package.
}
