% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{screen_cpp_MMLE_binomial}
\alias{screen_cpp_MMLE_binomial}
\title{Using MMLE Instead of MML in Step 2 of Sure Independence Screening for Binomial family}
\usage{
screen_cpp_MMLE_binomial(
  x,
  y,
  offset,
  num_keep,
  square = FALSE,
  main_effect = FALSE
)
}
\arguments{
\item{x}{a n-by-p matrix of main effects, with i.i.d rows, and each row represents a vector of observations of p main-effects}

\item{y}{a response vector of size \code{n}. \code{y} is a vector of continuous data in "gaussian" family; a vector of positive count data in "poisson" family; a vector of 0-1 binary data in "binomial" family. Notice that \code{y} should be a factor when using the "ordinal" family for ordinal response.}

\item{offset}{the linear predictions from Step 1.}

\item{num_keep}{the number of candidate interactions in Step 2. Default to be n.}

\item{square}{An indicator of whether squared effects should be considered in Step 1 (NOT Step 2!). square == TRUE if squared effects have been considered in Step 1, i.e., squared effects will NOT be considered in Step 2.}

\item{main_effect}{An indicator of whether main effects should also be screened. Default to be false.}
}
\value{
a three-column matrix representing the index pair of the selected interactions and its MMLE with the response vector given by the \code{offset}.
}
\description{
Using MMLE Instead of MML in Step 2 of Sure Independence Screening for Binomial family
}
