% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{screen_sparse_cpp}
\alias{screen_sparse_cpp}
\title{Sure Independence Screening in Step 2 for sparse design matrix}
\usage{
screen_sparse_cpp(x, y, num_keep, square = FALSE, main_effect = FALSE)
}
\arguments{
\item{x}{a n-by-p sparse matrix of main effects}

\item{y}{a vector of length n. In sprinter, y is the residual from step 1}

\item{num_keep}{the number of candidate interactions in Step 2. Default to be n / [log n]}

\item{square}{An indicator of whether squared effects should be considered in Step 1 (NOT Step 2!). square == TRUE if squared effects have been considered in Step 1, i.e., squared effects will NOT be considered in Step 2.}

\item{main_effect}{An indicator of whether main effects should also be screened. Default to be false. The functionality of main_effect = true is not used in sprinter, but for SIS_lasso.}
}
\value{
an matrix of 3 columns, representing the index pair of the selected interactions, and the corresponding absolute correlation with the residual.
}
\description{
Sure Independence Screening in Step 2 for sparse design matrix
}
