\name{sprintr-package}
\alias{sprintr-package}
\alias{sprintr}
\docType{package}
\title{
\packageTitle{sprintr}
}
\description{
\packageDescription{sprintr}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{sprintr}
\packageIndices{sprintr}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{sprintr}

Maintainer: \packageMaintainer{sprintr}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
