% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jira_api.R
\name{get_board_details}
\alias{get_board_details}
\title{Get details on a specific board}
\usage{
get_board_details(board_id)
}
\arguments{
\item{board_id}{ID of sprint board to retrieve}
}
\value{
dataframe
}
\description{
Fetches detailed information on a specific board.
}
\examples{
NULL
}
