% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jira_api.R
\name{get_issue}
\alias{get_issue}
\title{Title}
\usage{
get_issue(issue_key, full_response = FALSE)
}
\arguments{
\item{issue_key}{Key of issue to retrieve}

\item{full_response}{Return raw list of fields}
}
\value{
dataframe
}
\description{
Title
}
\examples{
NULL
}
