% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jira_api.R
\name{get_sprint_report}
\alias{get_sprint_report}
\title{Get sprint report data}
\usage{
get_sprint_report(sprint_id)
}
\arguments{
\item{sprint_id}{ID of sprint to retrieve}
}
\value{
dataframe
}
\description{
The Jira sprint report is a useful snapshot of issues completed, not completed,
and changes to sprint scope. While the web UI does not provide a way to
export this data in a workable format, the API can be used to get at the
raw data.
}
\examples{
NULL
}
