% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_sprint_report_detail}
\alias{get_sprint_report_detail}
\title{Get detailed sprint report information.}
\usage{
get_sprint_report_detail(sprint_id)
}
\arguments{
\item{sprint_id}{ID of the sprint to pull issue-level details on}
}
\value{
A dataframe
}
\description{
Get the issues associated with each of the major categories in a
Jira sprint report. Currently retrieves all fields of all issues, which
can be slow.
}
\examples{
NULL
}
