% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jira_api.R
\name{jira_api_post}
\alias{jira_api_post}
\title{Execute a POST request against the Jira API}
\usage{
jira_api_post(path, post_data)
}
\arguments{
\item{path}{API endpoint to retrieve}

\item{post_data}{List of data to post to API}
}
\value{
A jira_api object
}
\description{
Calls bare API end points and returns a wrapped \code{httr} response object
to a higher level function for presentation.
}
\examples{
NULL
}
