% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jira_api.R
\name{parse_issue}
\alias{parse_issue}
\title{Parse an issue.}
\usage{
parse_issue(response)
}
\arguments{
\item{response}{HTTR response}
}
\value{
dataframe
}
\description{
EXPERIMENTAL: Parse an issue reponse into a dataframe.
}
\examples{
NULL
}
