% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.sprinter.R
\name{plot.sprinter}
\alias{plot.sprinter}
\title{Plot function of sprinter fit}
\usage{
\method{plot}{sprinter}(fit, newdata, index = NULL, which = NULL,
  rugplot = TRUE, grid_length = 100, ...)
}
\arguments{
\item{fit}{The \code{sprinter} fit object.}

\item{newdata}{A new design matrix of all the main effects.}

\item{index}{Indices pair of the two tuning parameters at which the plot is produced. The default value is \code{c(1, length(fit$lambda))}.}

\item{which}{Indices pair of the interaction to plot. The default value is (1, 2).}

\item{rugplot}{If \code{TRUE} (default), adds a rugplot showing the values of x at the bottom of each fitted function plot.}

\item{grid_length}{The number of points to evaluate the estimated function at.}

\item{...}{Optional graphical parameters in the function \code{plot}.}
}
\description{
This function produces plots of the predicted values for specified main effects at a given tuning parameter values pair.
}
\details{
For each given query of interaction (j, k), this function produces four plots. The top two panels are the dependence of the predicted value only on the two main effects. The bottom left panel shows the dependence of the predicted value on the interaction x_j * x_k. And the bottom right panel shows the dependence of the predicted value on both the main effects and their interaction. Adopted from the function \code{plot.rgam} from package \code{relgam} by Kenneth Tay and Robert Tibshirani.
}
\examples{
set.seed(123)
n <- 100
p <- 100
# dense input
x <- matrix(rnorm(n * p), n, p)
y <- x[, 1] - 2 * x[, 2] + 3 * x[, 1] * x[, 3] - 4 * x[, 4] * x[, 5] + rnorm(n)
fit <- sprinter(x = x, y = y)

plot(fit, newdata = x, index = c(1, 50), which = c(4, 5))

# plot the fit from cv.sprinter
fit.cv <- cv.sprinter(x = x, y = y)
plot(fit.cv$fit, newdata = x, index = c(fit.cv$num_keep_best, fit.cv$lam_best))
}
