% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.sprinter.R
\name{print.sprinter}
\alias{print.sprinter}
\title{Print a summary of the sprinter fit}
\usage{
\method{print}{sprinter}(fit, which = 1, digits = max(3, getOption("digits")
  - 3), ...)
}
\arguments{
\item{fit}{A \code{sprinter} object.}

\item{which}{Which tuning paraemter of Step 2 to print. Default is 1.}

\item{digits}{Significant digits in printout.}

\item{...}{Additional print arguments.}
}
\description{
Print a summary of the sprinter fit at each step along the path of tuning parameters used in Step 3, for any given tuning parameter at Step 2.
}
\details{
The function produces a three-column matrix with tuning parameter values (in Step 3), number of nonzero main effects, and the number of nonzero interactions.
Adopted from the function \code{print.rgam} from package \code{relgam} by Kenneth Tay and Robert Tibshirani.
}
\examples{
set.seed(123)
n <- 100
p <- 100
x <- matrix(rnorm(n * p), n, p)
y <- x[, 1] - 2 * x[, 2] + 3 * x[, 1] * x[, 3] - 4 * x[, 4] * x[, 5] + rnorm(n)
fit <- sprinter(x = x, y = y)

print(fit)

# print the fit from cv.sprinter
fit.cv <- cv.sprinter(x = x, y = y)
print(fit.cv$fit, which = fit.cv$num_keep_best)

}
\seealso{
\code{\link{sprinter}}.
}
